update: use standard multioutputs loaders; update tests
