+++ base.py
@@ -8,6 +8,7 @@ import numpy as np
 import pandas as pd
 
 from typing import Tuple, List, Union
+from natsort import natsorted
 from sklearn.preprocessing import StandardScaler
 from loaders import CsvReader
 
@@ -17,7 +18,7 @@ console = logging.StreamHandler(sys.stdout)
 console.setLevel(logging.DEBUG)
 formater = logging.Formatter("%(name)-13s: %(levelname)-8s %(message)s")
 console.setFormatter(formater)
-logging.getLogger(__name__).addHandler(console)
+logging.getLogger("datamodeler").addHandler(console)
 
 # TODO: add weighting to the model
 # TODO: this should go into a metrics function?
@@ -203,10 +204,44 @@ class BaseModel(abc.ABC):
 
         pickle.dump(self.model, open(filename, "wb"))
 
-    def load_model(self, filename: str, scale_data: bool = False):
+    def load_model(
+        self, filename: str, scale_data: bool = False, separate_models: bool = False
+    ):
 
+        self.separate_models = separate_models
         self.scale_data = scale_data
-        self.model = pickle.load(open(filename, "rb"))
+
+        if scale_data:
+            if not self.separate_models:
+                path_name = str(pathlib.Path(filename).parent)
+            else:
+                path_name = filename
+            self.xscalar = pickle.load(
+                open(os.path.join(path_name, "xscalar.pkl"), "rb")
+            )
+            self.yscalar = pickle.load(
+                open(os.path.join(path_name, "yscalar.pkl"), "rb")
+            )
+        if separate_models:
+            self._load_multimodels(filename, scale_data)
+        else:
+            if not any([s in filename for s in [".pkl", ".pickle"]]):
+                filename += ".pkl"
+            self.model = pickle.load(open(filename, "rb"))
+
+    def _load_multimodels(self, filename: str, scale_data: bool):
+
+        all_models = os.listdir(filename)
+        all_models = natsorted(all_models)
+        if self.scale_data:
+            all_models = all_models[:-2]
+        num_models = len(all_models)
+        models = []
+        for i in range(num_models):
+            models.append(
+                pickle.load(open(os.path.join(filename, all_models[i]), "rb"))
+            )
+        self.models = models
 
     def evaluate(self, test_data: np.ndarray):
