+++ conf/config.yaml
@@ -1,4 +1,4 @@
 defaults:
-  - data: quanser-log.yaml
+  - data: cartpole_st1_at.yaml
   - model: SVR.yaml
-  - simulator: quanser-log.yaml
+  - simulator: gboost_cartpole.yaml
